Move cabinet wrapper to its own file
